Auto applmk1/requirements.txt
undetected-chromedriver==3.5.5
selenium==4.22.0
python-dotenv==1.0.1
PyYAML==6.0.2
Jinja2==3.1.4
pandas==2.2.2
